moved BioInput and UrlInput into own files, made EditDevForm default export
