+++ frontend/src/components/formInputs/BioInput/BioInput.js
@@ -1,0 +1,9 @@
+import React from "react"
+
+import BasicInput from "../BasicInput/BasicInput"
+
+const BioInput = props => (
+  <BasicInput componentName="bio" label="bio" type="textarea" {...props} />
+)
+
+export default BioInput
